Use dedicated projects API
